backend/requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
numpy
pdfplumber
python-dotenv
